Create the crud methods in controllers and repositories

+++ EFCoreDatabaseFirstSample/Models/BookAuthors.cs
@@ -5,8 +5,8 @@ namespace EFCoreDatabaseFirstSample.Models
 {
     public partial class BookAuthors
     {
-        public long BookId { get; set; }
-        public long AuthorId { get; set; }
+        public int BookId { get; set; }
+        public int AuthorId { get; set; }
 
         public virtual Author Author { get; set; }
         public virtual Book Book { get; set; }
